backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "deep-sort-realtime==1.3.2",
    "distro==1.9.0",
    "fastapi==0.115.12",
    "filelock==3.18.0",
    "fonttools==4.57.0",
    "fsspec==2025.3.2",
    "greenlet==3.2.1",
    "groq==0.23.0",
    "h11==0.14.0",
    "httpcore==1.0.8",
    "httpx==0.28.1",
    "idna==3.10",
    "jinja2==3.1.6",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "kiwisolver==1.4.8",
    "langchain==0.3.24",
    "langchain-core==0.3.55",
    "langchain-groq==0.3.2",
    "langchain-text-splitters==0.3.8",
    "langsmith==0.3.33",
    "markupsafe==3.0.2",
    "matplotlib==3.10.1",
    "mpmath==1.3.0",
    "networkx==3.4.2",
    "numpy==2.1.1",
    "opencv-python==4.11.0.86",
    "orjson==3.10.16",
    "packaging==24.2",
    "pandas==2.2.3",
    "pillow==11.2.1",
    "psutil==7.0.0",
    "py-cpuinfo==9.0.0",
    "pydantic==2.11.3",
    "pydantic-core==2.33.1",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "scipy==1.15.2",
    "seaborn==0.13.2",
    "setuptools==79.0.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.40",
    "sse-starlette==2.3.3",
    "starlette==0.46.2",
    "sympy==1.13.3",
    "tenacity==9.1.2",
    "torch==2.7.0",
    "torchvision==0.22.0",
    "tqdm==4.67.1",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "tzdata==2025.2",
    "ultralytics==8.3.115",
    "ultralytics-thop==2.0.14",
    "urllib3==2.4.0",
    "uvicorn[standard]==0.34.2",
    "zstandard==0.23.0",
]
